+++ test/updateSettings.js
@@ -1,0 +1,5 @@
+try {
+    require('../src/updateSettings.js')()
+} catch (err) {
+    console.error(err)
+}